backend/petpal/requirements.txt
asgiref==3.7.2
delocate==0.10.4
Django==4.2.6
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
networkx==3.1
packaging==23.1
Pillow==10.0.1
plotly==5.17.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytz==2023.3.post1
sqlparse==0.4.4
tenacity==8.2.3
typing_extensions==4.8.0
